Добавлено приведение типа

+++ libs/compiler/parser_expr.c
@@ -56,7 +56,7 @@ void must_be(parser *const prs, const token_t what, const error_t num)
 void applid(parser *const prs)
 {
 	prs->lastid = repr_get_reference(prs->sx, prs->lxr->repr);
-	if (prs->lastid == ITEM_MAX)
+	if (prs->lastid == (int)ITEM_MAX)
 	{
 		char buffer[MAXSTRINGL];
 		repr_get_name(prs->sx, REPRTAB_POS, buffer);
